project-b-enterprise-rag/requirements.txt
# Project B — hybrid RAG + rerank + guardrails
python-dotenv>=1.0.0
pypdf>=4.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
langchain-text-splitters>=0.2.0,<1
langchain-core>=0.3.0,<1
langchain>=0.3.0,<1
langchain-openai>=0.2.0,<1
langgraph>=0.2.40,<1
httpx>=0.27.0
openai>=1.0.0
rank-bm25>=0.2.2
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
prometheus-client>=0.20.0
SQLAlchemy>=2.0.0
websockets>=12.0
langfuse<3
